+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## TODO
+## 0.1.2
+
+### Added
+
+- Added `GameParameters` record which holds the constructor parameters for game classes.
 
 ### Changed
 
-- The game class generator now generates the game constructor and the static `Construct` method.
+- The game class generator now generates the static `Construct` method.
 - `Base` classes are no longer required. The game instance can now be accessed using the `Game.Self` property.
 
 ### Fixed